# docs/api/openapi.yaml
...
components:
  schemas:
...
    CheckinIn:
      properties:
        date:
          format: date
          title: Date
          type: string
        items:
          items:
            type: string
          title: Items
          type: array
      required:
      - date
      title: CheckinIn
      type: object
    CheckinOut:
      properties:
        date:
          format: date
          title: Date
          type: string
        items_json:
          items:
            type: string
          title: Items Json
          type: array
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - date
      - items_json
      - updated_at
      title: CheckinOut
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    PairMergeIn:
      description: 把相邻两个段落对合并成一个。``with_key`` 必须与 ``pair_key`` 相邻（F18）。
      properties:
        pair_key:
          title: Pair Key
          type: string
        with_key:
          title: With Key
          type: string
      required:
      - pair_key
      - with_key
      title: PairMergeIn
      type: object
    PairOut:
      properties:
        en:
          title: En
          type: string
        loc_page:
          anyOf:
          - type: integer
          - type: 'null'
          title: Loc Page
        manually_edited:
          title: Manually Edited
          type: boolean
        needs_review:
          title: Needs Review
          type: boolean
        pair_key:
          title: Pair Key
          type: string
        seq:
          title: Seq
          type: integer
        updated_at:
          format: date-time
          title: Updated At
          type: string
        zh:
          title: Zh
          type: string
      required:
      - pair_key
      - seq
      - zh
      - en
      - needs_review
      - manually_edited
      - updated_at
      title: PairOut
      type: object
    PairSplitIn:
      description: 把一个段落对拆成两个。两半中英文都由客户端算好（F18 结构化编辑）。
      properties:
        en_a:
          minLength: 1
          title: En A
          type: string
        en_b:
          minLength: 1
          title: En B
          type: string
        pair_key:
          title: Pair Key
          type: string
        zh_a:
          minLength: 1
          title: Zh A
          type: string
        zh_b:
          minLength: 1
          title: Zh B
          type: string
      required:
      - pair_key
      - zh_a
      - en_a
      - zh_b
      - en_b
      title: PairSplitIn
      type: object
...
    PieceOut:
      properties:
        align_mode:
          enum:
          - llm
          - regular
          title: Align Mode
          type: string
        id:
          title: Id
          type: string
        last_pos:
          title: Last Pos
          type: integer
        pair_count:
          default: 0
          title: Pair Count
          type: integer
        project_id:
          title: Project Id
          type: string
        recited:
          title: Recited
          type: boolean
        sort_order:
          title: Sort Order
          type: integer
        title:
          title: Title
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - id
      - project_id
      - title
      - align_mode
      - recited
      - last_pos
      - sort_order
      - updated_at
      title: PieceOut
      type: object
    PieceProgressIn:
      description: 背诵舱进度。客户端可离线写，联网后随写队列上传。
      properties:
        client_ts:
          format: date-time
          title: Client Ts
          type: string
        last_pos:
          minimum: 0.0
          title: Last Pos
          type: integer
        recited:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Recited
      required:
      - last_pos
      - client_ts
      title: PieceProgressIn
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: '学习工作台 Strayt 服务端。


    单用户免账号：所有 `/api/v1` 端点需 `Authorization: Bearer <访问令牌>`。

    唯一豁免是 `/health`（存活探针）。


    错误体统一为 `{ "error": { "code", "message", "detail?" } }`，`code` 是稳定契约，客户端据此分支。'
  title: Strayt 学习工作台
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/checkins:
    put:
      operationId: put_checkin_api_v1_checkins_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckinIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckinOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 打卡（同日覆盖）
      tags:
      - 背诵舱 / 打卡
  /api/v1/checkins/{day}:
    get:
      operationId: get_checkin_api_v1_checkins__day__get
      parameters:
      - in: path
        name: day
        required: true
        schema:
          format: date
          title: Day
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckinOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 查某天打卡
      tags:
      - 背诵舱 / 打卡
...
  /api/v1/projects/{project_id}/pieces:
    get:
      operationId: list_pieces_api_v1_projects__project_id__pieces_get
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PieceOut'
                title: Response List Pieces Api V1 Projects  Project Id  Pieces Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 篇目列表
      tags:
      - 背诵舱 / 打卡
  /api/v1/projects/{project_id}/pieces/{piece_id}/pairs:
    get:
      operationId: list_pairs_api_v1_projects__project_id__pieces__piece_id__pairs_get
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      - in: path
        name: piece_id
        required: true
        schema:
          title: Piece Id
          type: string
      - description: 续传起点
        in: query
        name: from_seq
        required: false
        schema:
          default: 0
          description: 续传起点
          minimum: 0
          title: From Seq
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 500
          maximum: 2000
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PairOut'
                title: Response List Pairs Api V1 Projects  Project Id  Pieces  Piece Id  Pairs Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 取段落对（背诵舱数据源）
      tags:
      - 背诵舱 / 打卡
  /api/v1/projects/{project_id}/pieces/{piece_id}/pairs/merge:
    post:
      description: '把相邻两个段落对合并成一个（中文直接拼接，英文用空格连接）。


        F18 结构化编辑：对着背诵舱里断开的对句手动合并。新对拿**新的 ``pair_key``**，

        标 ``manually_edited``，覆盖重跑不冲。'
      operationId: merge_pair_api_v1_projects__project_id__pieces__piece_id__pairs_merge_post
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      - in: path
        name: piece_id
        required: true
        schema:
          title: Piece Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairMergeIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PairOut'
                title: Response Merge Pair Api V1 Projects  Project Id  Pieces  Piece Id  Pairs Merge
                  Post
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 合并段落对（相邻两个合成一个）
      tags:
      - 背诵舱 / 打卡
  /api/v1/projects/{project_id}/pieces/{piece_id}/pairs/split:
    post:
      description: '把一个 ``(zh, en)`` 段落对按客户端算好的两半内容拆成两个新对。


        F18 结构化编辑：两句/两段并列在一个对里时手动拆开。拆的结果是**两个新的

        ``pair_key``**（稳定 key 算法与 AGENTS §7 的「人工内容与 AI 内容分区」一致），

        原对删除。两个新对都标 ``manually_edited``，覆盖重跑时不会被冲掉。'
      operationId: split_pair_api_v1_projects__project_id__pieces__piece_id__pairs_split_post
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      - in: path
        name: piece_id
        required: true
        schema:
          title: Piece Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairSplitIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PairOut'
                title: Response Split Pair Api V1 Projects  Project Id  Pieces  Piece Id  Pairs Split
                  Post
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 拆分段落对（拆成两个）
      tags:
      - 背诵舱 / 打卡
  /api/v1/projects/{project_id}/pieces/{piece_id}/progress:
    put:
      operationId: put_progress_api_v1_projects__project_id__pieces__piece_id__progress_put
      parameters:
      - in: path
        name: project_id
        required: true
        schema:
          title: Project Id
          type: string
      - in: path
        name: piece_id
        required: true
        schema:
          title: Piece Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PieceProgressIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PieceOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 上报背诵进度
      tags:
      - 背诵舱 / 打卡
...